type is set by createVariable, added extra test
