+++ include/csound_type_system.h
@@ -43,7 +43,8 @@ extern "C" {
     char* varTypeName;
     char* varDescription;
     int32_t argtype; // used to denote if allowed as in-arg, out-arg, or both
-    struct csvariable* (*createVariable)(void *cs, void *p, struct insds *ctx);
+    struct csvariable* (*createVariable)(void *cs, const struct cstype *p,
+                                         struct insds *ctx);
     void (*copyValue)(CSOUND* csound, const struct cstype* cstype, void* dest,
                       const void* src, struct insds *ctx);
     void (*freeVariableMemory)(void* csound, void* varMem);
